Correction for little oopsi ...
The bounding boxes were actually created based on face vertices rather than the segmentation as stated before.
This is sort of a bug in the postprocessing script. However, the trainings were done with the dataset which is
for download so it works just fine and is left it as is. Only need to fix the docu.